shanghai-house-price-analysis-system/backend/requirements.txt
Django==5.0.14
djangorestframework==3.17.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
